Refactored TabBarDelegate in external file
